fix type bundling
