+++ tsconfig.json
@@ -20,5 +20,5 @@
     "jsx": "react"
   },
   "include": ["src"],
-  "exclude": ["node_modules", "dist", "example", "rollup.config.js"]
+  "exclude": ["node_modules", "dist", "example", "rollup.config.js", "tests"]
 }
